refactor: eliminate services/oauth/ wrapper — absorb into bricks
Delete the 762-line services/oauth/OAuthService thin wrapper that violated
the LEGO architecture (bricks ARE the services layer). Changes:

- Add @rpc_expose to OAuthCredentialService's 6 public methods directly
- Move mcp_connect() (~300 lines) from OAuthService to MCPService
- Update factory wiring: create OAuthCredentialService directly from brick
- Update nexus_fs routing: use _SERVICE_ALIASES for oauth_* → brick methods
- Update mount_service: call revoke_credential() instead of oauth_revoke_credential()
- Update OAuthProtocol to use unprefixed method names
- Remove 3 stale import-linter ratchet entries for services.oauth
- Update all tests (oauth, smoke, protocol compliance, silent swallower)

Net: -439 lines, one fewer architectural violation.
