+++ pyproject.toml
@@ -941,7 +941,6 @@ ignore_imports = [
     "nexus.core.nexus_fs -> nexus.bricks.parsers.registry",
     "nexus.core.nexus_fs -> nexus.bricks.parsers.providers.registry",
     # --- services importing bricks ---
-    "nexus.services.oauth.oauth_service -> nexus.bricks.auth.oauth.credential_service",
     "nexus.services.workspace_rpc_service -> nexus.bricks.workspace.workspace_registry",
     # graph_search_service imports BaseSearchResult from bricks.search.results (Issue #2663)
     "nexus.services.search.graph_search_service -> nexus.bricks.search.results",
