--- src/nexus/services/oauth/__init__.py
@@ -1,8 +1,0 @@
-"""OAuth service domain -- BRICK tier.
-
-Canonical location for OAuth credential management.
-"""
-
-from nexus.services.oauth.oauth_service import OAuthService
-
-__all__ = ["OAuthService"]
